requirements.txt
uvicorn==0.30.6
fastapi==0.114.1
pydantic==2.9.1
google-cloud-storage==2.18.2
google-cloud-discoveryengine==0.13.4
